fix(controller): update import paths after package restructure
